Implement support for nested groups in nslcd
This differs from the code provided by Steve Hill in that it avoids
(recursively) performing parallel LDAP searches by queueing groups and check
for extra members per queued group (in the forward lookup) or check for extra
parents (for the user to groups lookup).

For the reverse lookup handling the NSLCD_HANDLE macro could no longer be used
because extra care should be taken to free the sets before returning and two
search phases are needed.