Define a NSLCD_ACTION_USERMOD request
The modification can either be requested by root or by the user itself.

Modifications by the user should be done by connecting to the LDAP server with
the user-supplied credentials. It is expected that access controls in the LDAP
server prevent unwanted modifications. The nslcd process is expected to check
whether supplied values are sensible.