Use clock_gettime() instead of gettimeofday()
This avoids problems with system clock changes (though there are some
safeguards in place to avoid waiting too long on clock changes).

Thanks to John Sullivan for pointing this out.

We can't easily use CLOCK_MONOTONIC_RAW or CLOCK_MONOTONIC_COARSE even
on platforms that define the clock because we can get runtime errors.
CLOCK_MONOTONIC seems to work on all tested platforms that provide it.