Portability fixes to environment tests
This mostly tries to reduce the influences of the test environment
(local users and groups) on the tests. This uses another username
(vsefcovic) in the PAM tests instead of the user arthur to avoid clashes
with existing users.

The PAM tests are skipped if passwd claims that it cannot modify LDAP
passwords (for FreeBSD).