update shadow.c to resolve pwdLastSet issue
we read the date into the buffer to the specified length to get it to the Unix time  (i.e. seconds) from its AD value of nanoseconds, then convert it to days for shadow. If we use date rather than buffer we end up trying to convert the original nanosecond value.