Fix formatting of size_t values
In several places the code used a %d format to print a size_t variable.
On amd64 at least size_t is an unsigned long, so use %lu instead.

An alternative would be to use %ud for size_t and %zd fo ssize_t but not
all platforms seem to support that formatter.