Implement disable_enumeration
If this option is present, functions which cause all user/group
entries to be loaded (getpwent(), getgrent()) from the
directory will not succeed in doing so. This can dramatically
reduce ldap server load in situations where there are a great
number of users and/or groups. Applications that
depend on being able to sequentially read all users and/or
groups may fail to operate correctly. This option is not recommended
for most configurations.