Loosen up file existence check
This changes the check (for configuration options that specify file
names) to just check that the specified path is readable instead of
ensisting that it points to a file.

This allows tls_randfile to point to /dev/urandom (a character device)
or a pipe. This fixes 6779a51.

This also applies the same check to the krb5_ccname option.

Thanks to Patrick McLean for pointing this out.