Query ethernet addresses in compact and long format
This ensures that when querying the address 0:18:8a:54:1a:8b both that
format and 00:18:8a:54:1a:8b is searched for in LDAP.

This was triggerred by the fact that ether_ntoa() on FreeBSD returns the
long format while glibc uses the compact format.

Since we are no longer using the libc version of ether_ntoa() we can
also drop the compatibility implementation of ether_ntoa_r().