Reorganise PAM search var building functions
This moves the autzsearch_var_add(), autzsearch_vars_free(),
autzsearch_var_get() and do_autzsearches() functions to the top of the
file using more generic names and introduces search_vars_new() in
prepartion of other similar searches.

This also renames the remaining authzsearch functions to authz_search to
be consistent with the pam_authz_search option.