Create /var/run/nslcd/socket after dropping privileges
This is needed to avoid a problem where a call to initgroups() can
result in NSS lookups. If nscd is configured the mechanism to avoid
loopback lookups using nss_ldap_enablelookups will not work and cause
for delays on start-up.

Note that this changes ownership of the socket to the user running
nslcd.