Fix handling of the pam_authc_ppolicy option
Check the result of the BIND operation instead of that of the
ldap_result() call when pam_authc_ppolicy is set to "no".

This could have resulted in successful authentication if the BIND
operation to the LDAP server timed out and pam_authc_ppolicy was set to
"no" but should not result in successful authentication otherwise so it
is unlikely that setting pam_authc_ppolicy to "no" ever worked as
intended. The timeout also would have to occur on the BIND operation,
not on setting up the connection.

Fixes 31cd2cf