Don't force password change during tests
Change the configuration of the password policy in the test suite to not
set pwdMustChange to TRUE. Between OpenLDAP 2.4 and 2.5 the behaviour of
the LDAP server was changed to force a password change whenever the
administrator changed a user's password. This change ensures that the
old behaviour is maintained.

See https://bugs.openldap.org/show_bug.cgi?id=7084